+++ src/components/buttons/process-button-container.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+export const processButtonContainer = to => props => {
+  if (typeof to !== 'string') {
+    return <button {...props}>{props.children}</button>
+  }
+
+  if (to.includes('https://') || to.includes('http://') || to.includes('//')) {
+    return (
+      <a {...props} href={to}>
+        {props.children}
+      </a>
+    )
+  }
+
+  return (
+    <Link {...props} to={to}>
+      {props.children}
+    </Link>
+  )
+}